fix: вынес проверку на наличие объекта из хранилища в сервис
