+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -3,9 +3,9 @@ package ru.yandex.practicum.filmorate.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.storage.FilmStorage;
-import ru.yandex.practicum.filmorate.storage.UserStorage;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -15,7 +15,7 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class FilmService {
     private final FilmStorage filmStorage;
-    private final UserStorage userStorage;
+    private final UserService userService;
 
     public Film create(Film film) {
         log.info("Create film {}", film);
@@ -24,6 +24,7 @@ public class FilmService {
 
     public Film update(Film film) {
         log.info("Update film {}", film);
+        validateFindFilmById(film.getId());
         return filmStorage.update(film);
     }
 
@@ -34,15 +35,15 @@ public class FilmService {
 
     public void addLike(Integer id, Integer userId) {
         log.info("Add like film {} user {}", id, userId);
-        userStorage.validateFindUserById(userId);
-        filmStorage.validateFindFilmById(id);
+        userService.validateFindUserById(userId);
+        validateFindFilmById(id);
         filmStorage.addLike(id, userId);
     }
 
     public void deleteLike(Integer id, Integer userId) {
         log.info("Delete like film {} user {}", id, userId);
-        userStorage.validateFindUserById(userId);
-        filmStorage.validateFindFilmById(id);
+        userService.validateFindUserById(userId);
+        validateFindFilmById(id);
         filmStorage.deleteLike(id, userId);
 
     }
@@ -57,11 +58,18 @@ public class FilmService {
 
     public Film getFilmById(Integer id) {
         log.debug("Get film by id {}", id);
+        validateFindFilmById(id);
         return filmStorage.getFilmById(id);
     }
 
     public void resetId() {
         log.debug("Reset film storage id");
         filmStorage.resetId();
+    }
+
+    public void validateFindFilmById(Integer id) {
+        if (!filmStorage.existsFilmById(id)) {
+            throw new NotFoundException(String.format("Не найден фильм %d", id), id);
+        }
     }
 }
