+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -3,6 +3,7 @@ package ru.yandex.practicum.filmorate.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.UserStorage;
 
@@ -24,6 +25,7 @@ public class UserService {
 
     public User update(User user) {
         log.info("Update user {}", user);
+        validateFindUserById(user.getId());
         if (user.getName().isBlank()) {
             user.setName(user.getLogin());
         }
@@ -37,31 +39,45 @@ public class UserService {
 
     public void addFriend(Integer id, Integer friendId) {
         log.info("Add user {} friend {}", id, friendId);
+        validateFindUserById(id);
+        validateFindUserById(friendId);
         userStorage.addFriend(id, friendId);
     }
 
     public void deleteFriend(Integer id, Integer friendId) {
         log.info("Delete user {} friend {}", id, friendId);
+        validateFindUserById(id);
+        validateFindUserById(friendId);
         userStorage.deleteFriend(id, friendId);
     }
 
     public List<User> getFriends(Integer id) {
         log.debug("Get friends, user {}", id);
+        validateFindUserById(id);
         return userStorage.getFriends(id);
     }
 
     public List<User> getCommonFriends(Integer id, Integer otherId) {
         log.debug("User {} get mutual friend {}", id, otherId);
+        validateFindUserById(id);
+        validateFindUserById(otherId);
         return userStorage.getCommonFriends(id, otherId);
     }
 
     public User getUserById(Integer id) {
         log.debug("Get user by id {}", id);
+        validateFindUserById(id);
         return userStorage.getUserById(id);
     }
 
     public void resetId() {
         log.debug("Reset id users");
         userStorage.resetId();
+    }
+
+    public void validateFindUserById(Integer id) {
+        if (!userStorage.existsUserById(id)) {
+            throw new NotFoundException(String.format("Не найден пользователь с ид %d", id), id);
+        }
     }
 }
