+++ src/main/java/ru/yandex/practicum/filmorate/storage/impl/InMemoryFilmStorage.java
@@ -1,7 +1,6 @@
 package ru.yandex.practicum.filmorate.storage.impl;
 
 import org.springframework.stereotype.Component;
-import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.storage.FilmStorage;
 
@@ -22,7 +21,6 @@ public class InMemoryFilmStorage implements FilmStorage {
 
     @Override
     public Film update(Film film) {
-        validateFindFilmById(film.getId());
         storage.put(film.getId(), film);
         return film;
     }
@@ -34,15 +32,7 @@ public class InMemoryFilmStorage implements FilmStorage {
 
     @Override
     public Film getFilmById(Integer id) {
-        validateFindFilmById(id);
         return storage.get(id);
-    }
-
-    @Override
-    public void validateFindFilmById(Integer id) {
-        if (storage.get(id) == null) {
-            throw new NotFoundException(String.format("Не найден фильм %d", id), id);
-        }
     }
 
     @Override
@@ -58,6 +48,11 @@ public class InMemoryFilmStorage implements FilmStorage {
     @Override
     public void resetId() {
         id = 0;
+    }
+
+    @Override
+    public boolean existsFilmById(Integer id) {
+        return storage.containsKey(id);
     }
 
     private Integer generatedId() {
